Add Dynamic Activity
Added a fragment to the main menu that is displayed upon clicking a button. The fragment contains buttons that show the use of Intent Extra Key-Value pairs.\n Also handles orientation change and the recreation of the activity / persistance of fragments.\nAdded the activity hierarchy to the activities as well.