Add Web Services
Added an example of the usage of web services and threads to the sample application.
Uses the OpenWeatherMap API data for a weekly weather data display;
Follows the Developing Android Apps course lessons 1 and 2 from Udacity;
Uses a list view based on a TextView;
The list is placed on a fragment and allows for the reuse of the base layout;
Uses a background thread for downloading the data, comments are available throughout the code; and
Uses an API Key for OpenWeatherMap API, this can be obtained by creating an account with OpenWeatherMap API. This process is free up to a certain usage amount.