Add Database
Added an example of the implementation of a SQLite Database.
Given that this database is to be used within this application only, a content provider was not used;
Added the Database Contract class;
Added the Database Helper class;
Added the Database Activity class;
Added the CRUD actions to the database;
Try, Catch and Finally used in order to ensure no SQL errors pass uncaught;
Shares the layout with ListView; and
Needs to be commented.